Move TR3 lighting manipulation to randomizers
